requirements.txt
python-telegram-bot==20.3
gspread==5.8.0
google-auth==2.21.0
